requirements.txt
flask
python-dotenv
google-generativeai
google-genai
pillow
requests
flask-login
werkzeug
email_validator
Flask-Limiter
psycopg2-binary
clerk-backend-api
python-jose[cryptography]
